pyproject.toml
[tool.poetry]
name = "drf-course-project"
version = "0.1.0"
description = ""
authors = ["Anton <anton_sima@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"
django-celery-beat = "^2.8.1"
celery = "^5.5.3"
dotenv = "^0.9.9"
psycopg2 = "^2.9.10"
requests = "^2.32.4"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
